completed code to save rev_jobs.txt data and clear file
Along with Windows's System Scheduler and a .bat file, code was updated to allow rev_graph.py to run every Monday at 12:00am.

When the batch file is run, rev_graph.py will execute and save a plot figure and the rev job data from the previous week to a local directory. It will then truncate/clear the old rev_jobs.txt file, allowing it to be used for the new week's data.